+++ src/Request.php
@@ -13,4 +13,13 @@ interface Request {
 	 */
 	public function getParams();
 
+	/**
+	 * Associative array of headers to add to the request.
+	 * Each key is the name of a header, and each value is a string or array of strings representing the header field values.
+	 *
+	 * @since 0.3
+	 * @return array
+	 */
+	public function getHeaders();
+
 }