+++ src/SimpleRequest.php
@@ -20,17 +20,24 @@ class SimpleRequest implements Request {
 	private $params;
 
 	/**
+	 * @var array
+	 */
+	private $headers;
+
+	/**
 	 * @param string $action
 	 * @param array $params
+	 * @param array $headers
 	 *
 	 * @throws InvalidArgumentException
 	 */
-	public function __construct( $action, array $params = array() ) {
+	public function __construct( $action, array $params = array(), array $headers = array() ) {
 		if( !is_string( $action ) ) {
 			throw new InvalidArgumentException( '$action must be string' );
 		}
 		$this->action = $action;
 		$this->params = $params;
+		$this->headers = $headers;
 	}
 
 	/**
@@ -40,5 +47,14 @@ class SimpleRequest implements Request {
 	 */
 	public function getParams() {
 		return array_merge( array( 'action' => $this->action ) , $this->params );
+	}
+
+	/**
+	 * @return array
+	 *
+	 * @since 0.3
+	 */
+	public function getHeaders() {
+		return $this->headers;
 	}
 }