+++ RELEASENOTES.md
@@ -5,6 +5,7 @@ These are the release notes for the [mediawiki-api-base](README.md).
 * UsageExceptions can now contain the full api result array
 * No longer uses addwiki/guzzle-mediawiki-client
 * Now using "guzzlehttp/guzzle": "~5.0" ( From "guzzle/guzzle": "~3.2" )
+* Added getHeaders method to Request interface
 
 ## Version 0.2 (13 January 2015)